Move E2E formatter tests into E2E suite
