+++ test/unit/test_gerber/test_formatter.py
@@ -4,8 +4,6 @@ from dataclasses import dataclass
 from io import StringIO
 from typing import Any
 
-import pytest
-
 from pygerber.gerber.formatter import Formatter
 from pygerber.gerber.formatter.enums import (
     EmptyLineBeforePolaritySwitch,
@@ -21,9 +19,6 @@ from pygerber.gerber.formatter.enums import (
 from pygerber.gerber.formatter.options import Options
 from pygerber.gerber.parser.pyparsing.parser import Parser
 from test.unit.test_gerber.common import (
-    GERBER_ASSETS_INDEX,
-    Asset,
-    CaseGenerator,
     ConfigBase,
 )
 
@@ -41,67 +36,6 @@ common_case_generator_config = {
     "macro.*": Config(dpmm=100),
     "incomplete.*": Config(skip=True),
 }
-
-
-parametrize = CaseGenerator(
-    GERBER_ASSETS_INDEX,
-    {
-        "A64_OLinuXino_rev_G.*": Config(dpmm=40),
-        "flashes.*": Config(dpmm=40),
-        "flashes.00_circle+h_4_tbh.grb": Config(
-            xfail=True,
-            xfail_message="Should warn, no mechanism implemented yet.",
-        ),
-        "ucamco.4.9.1.*": Config(dpmm=100),
-        "ucamco.4.9.6.*": Config(dpmm=300),
-        "ucamco.4.10.4.9.*": Config(dpmm=50),
-        "ucamco.4.11.4.*": Config(dpmm=1),
-        "expressions.*": Config(as_expression=True),
-        **common_case_generator_config,
-    },
-    Config,
-).parametrize
-
-
-parametrize = CaseGenerator(
-    GERBER_ASSETS_INDEX,
-    {
-        "A64_OLinuXino_rev_G.*": Config(skip=True),
-        "A64_OLinuXino_rev_G.A64-OlinuXino_Rev_G-B_Cu.gbr": Config(skip=False),
-        "ATMEGA328-Motor-Board.*": Config(skip=True),
-        "ATMEGA328-Motor-Board.ATMEGA328_Motor_Board-B.Cu.gbl": Config(skip=False),
-        "expressions.*": Config(as_expression=True),
-        **common_case_generator_config,
-    },
-    Config,
-).parametrize
-
-
-@parametrize
-def test_formatter(asset: Asset, config: Config) -> None:
-    if config.skip:
-        pytest.skip(reason=config.skip_reason)
-
-    if config.xfail:
-        pytest.xfail(config.xfail_message)
-
-    source = asset.absolute_path.read_text()
-    parser = Parser()
-
-    ast = parser.parse(source)
-    output_buffer = StringIO()
-    Formatter().format(ast, output_buffer)
-
-    output_buffer.seek(0)
-    formatted_source = output_buffer.read()
-
-    formatted_ast = parser.parse(formatted_source)
-
-    if formatted_ast.model_dump_json(serialize_as_any=True) != ast.model_dump_json(
-        serialize_as_any=True
-    ):
-        msg = "Formatted AST is the same as the original AST."
-        raise AssertionError(msg)
 
 
 DONUT_MACRO_SOURCE = """%AMDonut*
